add argmax argmin
