+++ test/test_alpha101.py
@@ -1,0 +1,3 @@
+import torch
+import unittest
+import featurizer.functors.alpha101 as alpha
